April Bug Fix Suite 0.1.7
