+++ AprilBugFixSuite/BeeData.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Xna.Framework;
+
+namespace AprilBugFixSuite
+{
+	public class BeeData
+	{
+		public double angle;
+		public int ticks;
+		public int currentSprite;
+		public Vector2 pos;
+		public Vector2 dir;
+	}
+}
