+++ AprilBugFixSuite/assets/raven.fr-FR.txt
@@ -1,0 +1,128 @@
+﻿Le Corbeau
+Par Edgar Allan Poe (traduit par Charles Baudelaire)
+
+    « Une fois, sur le minuit lugubre, pendant que je méditais, faible et
+fatigué, sur maint précieux et curieux volume d’une doctrine oubliée,
+pendant que je donnais de la tête, presque assoupi, soudain il se fit un
+tapotement, comme de quelqu’un frappant doucement, frappant à la porte
+de ma chambre. « C’est quelque visiteur, — murmurai-je, — qui frappe à la
+porte de ma chambre ; ce n’est que cela, et rien de plus. »
+
+    Ah ! distinctement je me souviens que c’était dans le glacial décembre,
+et chaque tison brodait à son tour le plancher du reflet de son agonie.
+Ardemment je désirais le matin ; en vain m’étais-je efforcé de tirer de mes
+livres un sursis à ma tristesse, ma tristesse pour ma Lénore perdue, pour la
+précieuse et rayonnante fille que les anges nomment Lénore, — et qu’ici on
+ne nommera jamais plus.
+
+    Et le soyeux, triste et vague bruissement des rideaux pourprés me
+pénétrait, me remplissait de terreurs fantastiques, inconnues pour moi
+jusqu’à ce jour ; si bien qu’enfin, pour apaiser le battement de mon cœur, je
+me dressai, répétant : « C’est quelque visiteur qui sollicite l’entrée à la porte
+de ma chambre, quelque visiteur attardé sollicitant l’entrée à la porte de ma
+chambre ; — c’est cela même, et rien de plus. »
+
+    Mon âme en ce moment se sentit plus forte. N’hésitant donc pas plus
+longtemps : « Monsieur, — dis-je, — ou madame, en vérité j’implore votre
+pardon ; mais le fait est que je sommeillais, et vous êtes venu frapper si
+doucement, si faiblement vous êtes venu taper à la porte de ma chambre,
+qu’à peine étais-je certain de vous avoir entendu. » Et alors j’ouvris la porte
+toute grande ; — les ténèbres, et rien de plus !
+
+    Scrutant profondément ces ténèbres, je me tins longtemps plein
+d’étonnement, de crainte, de doute, rêvant des rêves qu’aucun mortel n’a
+jamais osé rêver ; mais le silence ne fut pas troublé, et l’immobilité ne donna
+aucun signe, et le seul mot proféré fut un nom chuchoté : « Lénore ! » —
+C’était moi qui le chuchotais, et un écho à son tour murmura ce mot :
+« Lénore ! » — Purement cela, et rien de plus.
+
+    Rentrant dans ma chambre, et sentant en moi toute mon âme
+incendiée, j’entendis bientôt un coup un peu plus fort que le premier.
+« Sûrement, — dis-je, — sûrement, il y a quelque chose aux jalousies de ma
+fenêtre ; voyons donc ce que c’est, et explorons ce mystère. Laissons mon
+cœur se calmer un instant, et explorons ce mystère ; — c’est le vent, et rien
+de plus. »
+
+    Je poussai alors le volet, et, avec un tumultueux battement d’ailes, entra
+un majestueux corbeau digne des anciens jours. Il ne fit pas la moindre
+révérence, il ne s’arrêta pas, il n’hésita pas une minute ; mais, avec la mine
+d’un lord ou d’une lady, il se percha au-dessus de la porte de ma chambre ; il
+se percha sur un buste de Pallas juste au-dessus de la porte de ma
+chambre ; — il se percha, s’installa, et rien de plus.
+
+    Alors cet oiseau d’ébène, par la gravité de son maintien et la sévérité de
+sa physionomie, induisant ma triste imagination à sourire : « Bien que ta
+tête, — lui dis-je, — soit sans huppe et sans cimier, tu n’es certes pas un
+poltron, lugubre et ancien corbeau, voyageur parti des rivages de la nuit.
+Dis-moi quel est ton nom seigneurial aux rivages de la Nuit plutonienne ! »
+Le corbeau dit : « Jamais plus ! »
+
+    Je fus émerveillé que ce disgracieux volatile entendît si facilement la
+parole, bien que sa réponse n’eût pas un bien grand sens et ne me fût pas
+d’un grand secours ; car nous devons convenir que jamais il ne fut donné à un
+homme vivant de voir un oiseau au-dessus de la porte de sa chambre, un
+oiseau ou une bête sur un buste sculpté au-dessus de la porte de sa
+chambre, se nommant d’un nom tel que Jamais plus !
+
+    Mais le corbeau, perché solitairement sur le buste placide, ne proféra
+que ce mot unique, comme si dans ce mot unique il répandait toute son âme.
+Il ne prononça rien de plus ; il ne remua pas une plume, — jusqu’à ce
+que je me prisse à murmurer faiblement : « D’autres amis se sont déjà
+envolés loin de moi ; vers le matin, lui aussi, il me quittera comme mes
+anciennes espérances déjà envolées. » L’oiseau dit alors : « Jamais plus ! »
+
+    Tressaillant au bruit de cette réponse jetée avec tant d’à-propos : « Sans
+doute, — dis-je, — ce qu’il prononce est tout son bagage de savoir, qu’il a
+pris chez quelque maître infortuné que le Malheur impitoyable a poursuivi
+ardemment, sans répit, jusqu’à ce que ses chansons n’eussent plus qu’un
+seul refrain, jusqu’à ce que le De profundis de son Espérance eût pris ce
+mélancolique refrain : Jamais, jamais plus !
+
+    Mais, le corbeau induisant encore toute ma triste âme à sourire, je
+roulai tout de suite un siège à coussins en face de l’oiseau et du buste et de
+la porte ; alors, m’enfonçant dans le velours, je m’appliquai à enchaîner les
+idées aux idées, cherchant ce que cet augural oiseau des anciens jours, ce
+que ce triste, disgracieux, sinistre, maigre et augural oiseau des anciens
+jours voulait faire entendre en croassant son Jamais plus !
+
+    Je me tenais ainsi, rêvant, conjecturant, mais n’adressant plus une
+syllabe à l’oiseau, dont les yeux ardents me brûlaient maintenant jusqu’au
+fond du cœur ; je cherchais à deviner cela, et plus encore, ma tête reposant
+à l’aise sur le velours du coussin que caressait la lumière de la lampe, ce
+velours violet caressé par la lumière de la lampe que sa tête, à Elle, ne
+pressera plus, — ah ! jamais plus !
+
+    Alors il me sembla que l’air s’épaississait, parfumé par un encensoir
+invisible que balançaient des séraphins dont les pas frôlaient le tapis de la
+chambre. « Infortuné ! — m’écriai-je, — ton Dieu t’a donné par ses anges, il
+t’a envoyé du répit, du répit et du népenthès dans tes ressouvenirs de
+Lénore ! Bois, oh ! bois ce bon népenthès, et oublie cette Lénore perdue ! »
+Le corbeau dit : « Jamais plus ! »
+
+    « Prophète ! — dis-je, — être de malheur ! oiseau ou démon, mais
+toujours prophète ! que tu sois un envoyé du Tentateur, ou que la tempête
+t’ait simplement échoué, naufragé, mais encore intrépide, sur cette terre
+déserte, ensorcelée, dans ce logis par l’Horreur hanté, — dis-moi
+sincèrement, je t’en supplie, existe-t-il, existe-t-il ici un baume de Judée ? Dis,
+dis, je t’en supplie ! » Le corbeau dit : « Jamais plus ! »
+
+    « Prophète ! — dis-je, — être de malheur ! oiseau ou démon ! toujours
+prophète ! par ce Ciel tendu sur nos têtes, par ce Dieu que tous deux nous
+adorons, dis à cette âme chargée de douleur si, dans le Paradis lointain, elle
+pourra embrasser une fille sainte que les anges nomment Lénore,
+embrasser une précieuse et rayonnante fille que les anges nomment
+Lénore. » Le corbeau dit : « Jamais plus ! »
+
+    « Que cette parole soit le signal de notre séparation, oiseau ou démon !
+— hurlai-je en me redressant. — Rentre dans la tempête, retourne au rivage
+de la Nuit plutonienne ; ne laisse pas ici une seule plume noire comme
+souvenir du mensonge que ton âme a proféré ; laisse ma solitude inviolée ;
+quitte ce buste au-dessus de ma porte ; arrache ton bec de mon cœur et
+précipite ton spectre loin de ma porte ! » Le corbeau dit : « Jamais plus ! »
+
+    Et le corbeau, immuable, est toujours installé, toujours installé sur le
+buste pâle de Pallas, juste au-dessus de la porte de ma chambre ; et ses
+yeux ont toute la semblance des yeux d’un démon qui rêve ; et la lumière de
+la lampe, en ruisselant sur lui, projette son ombre sur le plancher ; et mon
+âme, hors du cercle de cette ombre qui gît flottante sur le plancher, ne
+pourra plus s’élever, — jamais plus !
