+++ AprilBugFixSuite/i18n/fr.json
@@ -1,0 +1,51 @@
+{
+	// Tree screams
+	"tree-chop-1": "Ah ! Arrêtez !",
+	"tree-chop-2": "Oh, la cruauté !",
+	"tree-chop-3": "Ayez pitié !",
+	"tree-chop-4": "Aïe !",
+	"tree-chop-5": "S'il vous plaît, je vous en supplie !",
+	"tree-chop-6": "Oh, ça fait mal !",
+	"tree-chop-7": "Sauvez-moi, Yoba !",
+	"tree-fall-1": "Aaaaaaaah !",
+	"tree-fall-2": "Et tu, Brute ?",
+	"tree-fall-3": "Je suis une feuille au vent...",
+	"tree-fall-4": "Libertééééé !",
+
+	// Animal speeches
+	"animal-speech-1-1": "L'erreur est humaine...",
+	"animal-speech-1-2": "Le pardon est bovin.",
+	"animal-speech-2-1": "Si ça parle comme un canard...",
+	"animal-speech-2-2": "Alors tu es ivre. Les canards ne parlent pas.",
+	"animal-speech-3-1": "Pourquoi les canards sont toujours à l'heure ?",
+	"animal-speech-3-2": "Parce qu'ils sont dans l'étang.",
+	"animal-speech-4-1": "Que fait une vache quand elle réfléchit ?",
+	"animal-speech-4-2": "Du lait concentré.",
+	"animal-speech-5-1": "Quels fruits obtient-on en tirant sur la queue d'une vache ?",
+	"animal-speech-5-2": "Un meuh long.",
+	"animal-speech-6-1": "Combien faut-il de moutons pour tricoter un pull ?",
+	"animal-speech-6-2": "Les moutons ne tricotent pas.",
+	"animal-speech-7-1": "Pourquoi est-ce que les moutons aiment le chewing-gum ?",
+	"animal-speech-7-2": "Parce que c’est bon pour la laine.",
+	"animal-speech-8-1": "Pourquoi est-ce qu’on met tous les crocos en prison ?",
+	"animal-speech-8-2": "Parce que les crocos dealent.",
+	"animal-speech-9-1": "Que dit une vache quand elle prend une photo ?",
+	"animal-speech-9-2": "Ne bouse plus.",
+	"animal-speech-10-1": "Comment appelle-t-on les petits d’une oie ?",
+	"animal-speech-10-2": "Les noisettes.",
+
+	// GMCM
+	"GMCM.ModEnabled.Name": "Activer le Mod",
+	"GMCM.OnlyAprilFirst.Name": "Seulement le 1er avril",
+	"GMCM.EnableBees.Name": "Activer les abeilles",
+	"GMCM.EnableBackwards.Name": "Activer la marche inversée",
+	"GMCM.EnableAscii.Name": "Activer l'ASCII",
+	"GMCM.EnablePixelate.Name": "Activer la pixellisation",
+	"GMCM.EnableTreeScreams.Name": "Activer les cris des arbres",
+	"GMCM.EnableInventoryAvoid.Name": "Activer l'évitement d'inventaire",
+	"GMCM.EnableSlime.Name": "Activer le slime prismatique",
+	"GMCM.EnableRaven.Name": "Activer les corbeaux",
+	"GMCM.EnableGiants.Name": "Activer les géants",
+	"GMCM.EnableBuildingSwitch.Name": "Activer le changement d'emplacement des bâtiments",
+	"GMCM.EnableAnimalGreeting.Name": "Activer les salutations des animaux"
+}
