+++ AprilBugFixSuite/i18n/ru.json
@@ -1,0 +1,51 @@
+{
+	// Tree screams
+	"tree-chop-1": "Ай! Хватит!",
+	"tree-chop-2": "Как жестоко!",
+	"tree-chop-3": "Пощади!",
+	"tree-chop-4": "Ай!",
+	"tree-chop-5": "Пожалуйста, я тебя умоляю!",
+	"tree-chop-6": "А-а! Это больно!",
+	"tree-chop-7": "Спаси меня, Йоба!",
+	"tree-fall-1": "Ааааааа!",
+	"tree-fall-2": "И ты, Брут?",
+	"tree-fall-3": "Я лишь лепесток на ветру...",
+	"tree-fall-4": "Свобода!",
+
+	// Animal speeches
+	"animal-speech-1-1": "Человеку свойственно совершать ошибки...",
+	"animal-speech-1-2": "Чтобы простить скот.",
+	"animal-speech-2-1": "Если оно говорит как утка...",
+	"animal-speech-2-2": "То белочка уже пришла. Утки не разговаривают.",
+	"animal-speech-3-1": "Если оно говорит как утка...",
+	"animal-speech-3-2": "То белочка уже пришла. Утки не разговаривают.",
+	"animal-speech-4-1": "Почему у коров копыта, а не ноги?",
+	"animal-speech-4-2": "Потому что они дают молоко.",
+	"animal-speech-5-1": "Тебе известно о весёлых фермерах?",
+	"animal-speech-5-2": "Они жили семенами своих растений.",
+	"animal-speech-6-1": "Тебе известно про призовое пугало?",
+	"animal-speech-6-2": "Оно точно стояло не в своей тарелке.",
+	"animal-speech-7-1": "Они говорили: готовить сено сложно.",
+	"animal-speech-7-2": "Для меня всё очень просто.",
+	"animal-speech-8-1": "Они говорили: готовить сено сложно.",
+	"animal-speech-8-2": "Для меня всё очень просто.",
+	"animal-speech-9-1": "Пора на сеновал.",
+	"animal-speech-9-2": "Тяжела и неказиста жизнь у... тракториста.",
+	"animal-speech-10-1": "Пчёлы боятся меня. Я видел их истинные лица.",
+	"animal-speech-10-2": "Тела - продолжения злых лиц. А лица наполнены злобой.",
+
+	// GMCM
+	"GMCM.ModEnabled.Name": "Включить мод?",
+	"GMCM.OnlyAprilFirst.Name": "Только на первое апреля?",
+	"GMCM.EnableBees.Name": "Включить пчёл?",
+	"GMCM.EnableBackwards.Name": "Включить обратный режим?",
+	"GMCM.EnableAscii.Name": "Включить ASCii?",
+	"GMCM.EnablePixelate.Name": "Включить пикселизацию?",
+	"GMCM.EnableTreeScreams.Name": "Включить крики деревьев?",
+	"GMCM.EnableInventoryAvoid.Name": "Включить избегание инвентаря?",
+	"GMCM.EnableSlime.Name": "Включить слаймов?",
+	"GMCM.EnableRaven.Name": "Включить ворон?",
+	"GMCM.EnableGiants.Name": "Включить гигантов?",
+	"GMCM.EnableBuildingSwitch.Name": "Включить переключение построек?",
+	"GMCM.EnableAnimalGreeting.Name": "Включить приветствие животных?"
+}
