--- AprilFools/BeeData.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.Xna.Framework;
-
-namespace AprilFools
-{
-    public class BeeData
-    {
-        public double angle;
-        public int ticks;
-        public int currentSprite;
-        public Vector2 pos;
-        public Vector2 dir;
-    }
-}
