--- AprilFools/IGenericModConfigMenuApi.cs
@@ -1,48 +1,0 @@
-﻿using StardewModdingAPI;
-using System;
-
-namespace AprilFools
-{
-    /// <summary>The API which lets other mods add a config UI through Generic Mod Config Menu.</summary>
-    public interface IGenericModConfigMenuApi
-    {
-        /*********
-        ** Methods
-        *********/
-        /****
-        ** Must be called first
-        ****/
-        /// <summary>Register a mod whose config can be edited through the UI.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="reset">Reset the mod's config to its default values.</param>
-        /// <param name="save">Save the mod's current config to the <c>config.json</c> file.</param>
-        /// <param name="titleScreenOnly">Whether the options can only be edited from the title screen.</param>
-        /// <remarks>Each mod can only be registered once, unless it's deleted via <see cref="Unregister"/> before calling this again.</remarks>
-        void Register(IManifest mod, Action reset, Action save, bool titleScreenOnly = false);
-
-        /// <summary>Add a boolean option at the current position in the form.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="getValue">Get the current value from the mod config.</param>
-        /// <param name="setValue">Set a new value in the mod config.</param>
-        /// <param name="name">The label text to show in the form.</param>
-        /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
-        /// <param name="fieldId">The unique field ID for use with <see cref="OnFieldChanged"/>, or <c>null</c> to auto-generate a randomized ID.</param>
-        void AddBoolOption(IManifest mod, Func<bool> getValue, Action<bool> setValue, Func<string> name, Func<string> tooltip = null, string fieldId = null);
-
-        /// <summary>Add an integer option at the current position in the form.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="getValue">Get the current value from the mod config.</param>
-        /// <param name="setValue">Set a new value in the mod config.</param>
-        /// <param name="name">The label text to show in the form.</param>
-        /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
-        /// <param name="min">The minimum allowed value, or <c>null</c> to allow any.</param>
-        /// <param name="max">The maximum allowed value, or <c>null</c> to allow any.</param>
-        /// <param name="interval">The interval of values that can be selected.</param>
-        /// <param name="fieldId">The unique field ID for use with <see cref="OnFieldChanged"/>, or <c>null</c> to auto-generate a randomized ID.</param>
-        void AddNumberOption(IManifest mod, Func<int> getValue, Action<int> setValue, Func<string> name, Func<string> tooltip = null, int? min = null, int? max = null, int? interval = null, string fieldId = null);
-
-        /// <summary>Remove a mod from the config UI and delete all its options and pages.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        void Unregister(IManifest mod);
-    }
-}
