--- AprilFools/ModConfig.cs
@@ -1,22 +1,0 @@
-﻿
-using StardewModdingAPI;
-
-namespace AprilFools
-{
-    public class ModConfig
-    {
-        public bool EnableMod { get; set; } = true;
-        public bool RestrictToAprilFirst { get; set; } = true;
-        public bool BeesEnabled { get; set; } = true;
-        public bool BackwardsEnabled { get; set; } = true;
-        public bool AsciiEnabled { get; set; } = true;
-        public bool PixelateEnabled { get; set; } = true;
-        public bool TreeScreamEnabled { get; set; } = true;
-        public bool InventoryEnabled { get; set; } = true;
-        public bool RavenEnabled { get; set; } = true;
-        public bool SlimeEnabled { get; set; } = true;
-        public bool GiantEnabled { get; set; } = true;
-        public bool BuildingsEnabled { get; set; } = true;
-        public bool EnableAnimalTalk { get; set; } = true;
-    }
-}
