--- AprilFools/manifest.json
@@ -1,22 +1,0 @@
-﻿{
-  "Name": "April Fools",
-  "Author": "aedenthorn",
-  "Version": "0.1.6",
-  "Description": "April Fools.",
-  "UniqueID": "aedenthorn.AprilFools",
-  "EntryDll": "AprilFools.dll",
-  "MinimumApiVersion": "3.13.0",
-  "ModUpdater": {
-    "Repository": "StardewValleyMods",
-    "User": "aedenthorn",
-    "Directory": "_releases",
-    "ModFolder": "AprilFools"
-  },
-  "UpdateKeys": [ "Nexus:" ],
-  "Dependencies": [
-    {
-      "UniqueID": "Platonymous.ModUpdater",
-      "IsRequired": false
-    }
-  ]
-}
